Setting up search interface
