+++ .idea/misc.xml
@@ -9,6 +9,7 @@
         <entry key="..\:/Users/15169/AndroidStudioProjects/MercadoPago/app/src/main/res/layout/fragment_search.xml" value="0.23148148148148148" />
         <entry key="..\:/Users/15169/AndroidStudioProjects/MercadoPago/app/src/main/res/menu-v26/options_menu.xml" value="0.103125" />
         <entry key="..\:/Users/15169/AndroidStudioProjects/MercadoPago/app/src/main/res/menu/options_menu.xml" value="0.103125" />
+        <entry key="..\:/Users/15169/AndroidStudioProjects/MercadoPago/app/src/main/res/xml/searchable.xml" value="0.21910112359550563" />
       </map>
     </option>
   </component>
