+++ app/src/main/java/com/projects/mercadopago/uiControllers/activities/SearchResultsActivity.kt
@@ -1,0 +1,12 @@
+package com.projects.mercadopago.uiControllers.activities
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import com.projects.mercadopago.R
+
+class SearchResultsActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_search_results)
+    }
+}
