+++ app/src/main/res/values/strings.xml
@@ -5,4 +5,5 @@
     <string name="search_title">Search</string>
     <string name="filter_by_product">Filter by product</string>
     <string name="products">Products</string>
+    <string name="search_hint">Search MercadoPago</string>
 </resources>
